Expose `window::Mode` in `iced`
Although the Fullscreen API in the Web platform has some limitations, it
is still useful to be able to support fullscreen on the native side.
