+++ native/src/window.rs
@@ -1,8 +1,6 @@
 //! Build window-based GUI applications.
 mod event;
-mod mode;
 mod renderer;
 
 pub use event::Event;
-pub use mode::Mode;
 pub use renderer::{Renderer, Target};
