+++ src/settings.rs
@@ -1,43 +1,21 @@
 //! Configure your application.
+use crate::window;
 
 /// The settings of an application.
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Default)]
 pub struct Settings {
-    /// The [`Window`] settings.
+    /// The window settings.
     ///
     /// They will be ignored on the Web.
     ///
     /// [`Window`]: struct.Window.html
-    pub window: Window,
+    pub window: window::Settings,
 
     /// The bytes of the font that will be used by default.
     ///
     /// If `None` is provided, a default system font will be chosen.
     // TODO: Add `name` for web compatibility
     pub default_font: Option<&'static [u8]>,
-}
-
-/// The window settings of an application.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct Window {
-    /// The size of the window.
-    pub size: (u32, u32),
-
-    /// Whether the window should be resizable or not.
-    pub resizable: bool,
-
-    /// Whether the window should have a border, a title bar, etc. or not.
-    pub decorations: bool,
-}
-
-impl Default for Window {
-    fn default() -> Window {
-        Window {
-            size: (1024, 768),
-            resizable: true,
-            decorations: true,
-        }
-    }
 }
 
 #[cfg(not(target_arch = "wasm32"))]
