+++ src/window.rs
@@ -1,0 +1,6 @@
+//! Configure the window of your application in native platforms.
+mod mode;
+mod settings;
+
+pub use mode::Mode;
+pub use settings::Settings;
